requirements.txt
django
django-configurations
django-rest-framework
django-suit
gunicorn
honcho
